Khichuri

+++ Carnesia.Domain/CMS/ComboProducts/AddComboDetailsDTO.cs
@@ -11,7 +11,7 @@ namespace Carnesia.Domain.CMS.ComboProducts
         public string comboProductName { get; set; }
         public string sku { get; set; }
         public int productId { get; set; }
-        public int regularPrice { get; set; }
+        public decimal regularPrice { get; set; }
         public int comboPrice { get; set; }
         public string description { get; set; }
         public string startDate { get; set; }
